refactoring : #6 폴더에 클래스 분류 ### 추가 - domain   - InputNumbers   - Numbers   - TargetNumbers   - PitchingResult
### 이유
- 각 클래스 명에 따라 용도를 오해할 수 있을 것으로 판단하여 이동 시켰습니다.
