+++ app/src/main/java/edu/number/baseball/NumberBaseballPitching.java
@@ -1,5 +1,8 @@
 package edu.number.baseball;
 
+import edu.number.baseball.domain.Numbers;
+import edu.number.baseball.domain.PitchingResult;
+
 import java.util.List;
 
 public class NumberBaseballPitching {
