+++ app/src/main/java/edu/number/baseball/domain/TargetNumbers.java
@@ -1,4 +1,4 @@
-package edu.number.baseball;
+package edu.number.baseball.domain;
 
 import edu.number.baseball.randomnumber.RandomNumberGenerationRule;
 import lombok.Getter;
